Configure interactive Android host tests
