+++ haze/src/commonTest/kotlin/dev/chrisbanes/haze/InteractiveVisualEffectTest.kt
@@ -3,24 +3,16 @@
 
 package dev.chrisbanes.haze
 
-import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.drawscope.DrawScope
-import androidx.compose.ui.graphics.toPixelMap
 import androidx.compose.ui.input.pointer.PointerEvent
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.test.ExperimentalTestApi
-import androidx.compose.ui.test.assertHeightIsEqualTo
-import androidx.compose.ui.test.assertWidthIsEqualTo
-import androidx.compose.ui.test.captureToImage
 import androidx.compose.ui.test.click
 import androidx.compose.ui.test.onNodeWithTag
 import androidx.compose.ui.test.performTouchInput
@@ -109,35 +101,9 @@ class InteractiveVisualEffectTest : ContextTest() {
     assertThat(effect.cancelCalls).isEqualTo(1)
     assertThat(effect.pointerEvents.size).isEqualTo(eventCount)
   }
-
-  @Test
-  fun materialAndContentTransform_scalesFinalGroupWithoutChangingBounds() = runComposeUiTest {
-    val effect = RecordingInteractiveVisualEffect().apply {
-      transform = VisualEffectTransform(0.5f, 0.5f, Offset(50f, 50f))
-    }
-    setContent {
-      Box(Modifier.size(100.dp).background(Color.Black)) {
-        Box(
-          Modifier
-            .fillMaxSize()
-            .testTag("glass")
-            .hazeEffect {
-              drawContentBehind = true
-              visualEffect = effect
-            }
-            .background(Color.Red),
-        )
-      }
-    }
-
-    val pixels = onNodeWithTag("glass").captureToImage().toPixelMap()
-    assertThat(pixels[50, 50]).isEqualTo(Color.Red)
-    assertThat(pixels[5, 5]).isEqualTo(Color.Black)
-    onNodeWithTag("glass").assertWidthIsEqualTo(100.dp).assertHeightIsEqualTo(100.dp)
-  }
 }
 
-private class RecordingInteractiveVisualEffect : InteractiveVisualEffect {
+internal class RecordingInteractiveVisualEffect : InteractiveVisualEffect {
   var observes by mutableStateOf(false)
   var transform: VisualEffectTransform = VisualEffectTransform.Identity
   val pointerEvents = mutableListOf<PointerEvent>()
